REFACTOR/US-29: add fields to User and Project in accordance with initial database schema

+++ app/models/Client.scala
@@ -4,7 +4,7 @@ import play.api.libs.json.{Json, OWrites, Reads}
 import scala.collection.mutable.ArrayBuffer
 
 case class Client(
-  id: Int = Client.all.maxBy(_.id).id + 1,
+  id: BigInt = Client.all.maxBy(_.id).id + 1,
   name: String = "client " + (User.all.maxBy(_.id).id + 1)
 )
